refactor | code && delete comment lines
ignore lock files

add: clean command

remove innecesary lines

remove comment lines

remove innecesary code

add: transform.ts | allow to transform strings with colors defined

delete: helpers.ts && yarn.lock
