+++ src/index.ts
@@ -1,13 +1,1 @@
 export { println, print } from './print';
-
-
-// // Showing basic colors
-// println`{cyan}This is my other message{reset}`;
-// println`{green}success: {magenta}Intl-graphl{reset}`;
-
-// // TODO: Interpolation support
-// const lang: string = 'en';
-
-// println`{green}sucess: {white}plugin intl-graphql for {red}${2+2}{white} languages {yellow}"${lang}"{reset}`;
-
-// console.log("");
