+++ src/print.ts
@@ -1,29 +1,8 @@
 import { stdout } from "process";
 import { colors } from "./colors";
-import { transform } from "./helpers";
+import { transform } from "./transform";
 
-const regexToCheckDollarNotation = /\$[a-z]+/g;
 const regexToCheckBreackets = /{[a-z]+}/g;
-
-/* 
-  Will there 2 forms to show the colors.
-
-  The first way is with "$"
-  The second way is putting the color between "{}"
-
-  example:
-
-  1) print`$red There an error! $reset`
-  2) print`{red} There an error! {reset}`
-
-  This function check which way was used, to apply
-  the correct string clean.
-*/
-
-// Other way of represent the colors
-// function clearDollarSign() {
-
-// }
 
 function clearCurlyBrackets(entries: RegExpMatchArray | null) {
   if (!entries?.length) return;
@@ -31,26 +10,15 @@ function clearCurlyBrackets(entries: RegExpMatchArray | null) {
 }
 
 function checkEntryType(entry: string): string {
-  if (entry.match(regexToCheckBreackets)) {
-    const result: RegExpMatchArray | null = entry.match(regexToCheckBreackets);
-    let codes = findColor(clearCurlyBrackets(result));
-    
-    let final: string = transform({
-      result,
-      codes, 
-      content: entry
-    })
 
-    return final;
+  if (!entry.match(regexToCheckBreackets)) return entry;
 
-  } else if (entry.match(regexToCheckDollarNotation)) {
-    // TODO
-    console.log('Use $[color]')
-  } else {
-    throw new TypeError('This way is not valid!')
-  }
+  const result: RegExpMatchArray | null = entry.match(regexToCheckBreackets);
+  let codes = findColor(clearCurlyBrackets(result));
 
-  return "";
+  let final = transform({ result, codes, content: entry });
+
+  return final;
 }
 
 function findColor(entries: Array<string> | undefined) {
@@ -69,5 +37,5 @@ export function println(message: TemplateStringsArray, ...args: any[]): void {
 
 export function print(message: TemplateStringsArray, ...args: any[]): void {
   let result: string = mergeStrings(message, args);
-  stdout.write(checkEntryType(result)); 
+  stdout.write(checkEntryType(result));
 }
